gas: Restructures the energy modes in the electronically specific gas model to allow for relaxation between temperatures.
This is a large restructure of the thermodynamics in the electronically specific gas model which no longer treats
the raised energy states as a separate energy mode, and instead relies upon the summation of species energies
to get the energy stored. Instead, correct treatment of the electron energy in the first additional energy mode
allows for the transfer of energy between electrons and heavy particles using the Appleton and Bray model.

Currently, energies seem too high in the first energy mode for electron energies. Will revisit soon.