kinetics: Included energy mode relaxation in the kinetics of the electronically specific gas model.
This commit included a number of changes to connect the energy mode relaxation to the gas kinetics.
Electronic energy levels were included as a property in gas model. The function for energy relaxation
was moved from the gas file to the kinetics file. The electronic state solver was altered to allow
for dtChemSuggest to be supplied to the solver.