kinetics: Final changes to electronic state model to allow for non-binned modelling.
This commit resolves a number of issues encountered when trying to manipulate the previously implemented
electronics solver to model completely ungrouped eletronic energy levels. It also includes gas model
files for the two cases of either the selected binning strategy, or an un-binned model.

Editing of the kinetics file is required for binned or non-binned runs.