Eilmer: introduce OutFlowBC_SimpleFlux boundary condition.
This commit is just the house-keeping code needed to construct and apply the
boundary condition.  The actual computations details will (eventually) be held
in the BoundaryFluxEffect subsclass BFE_SimpleOutflowFlux.

This commit also renames OutFlowBC_Simple as OutFlowBC_SimpleExtrapolate but
retains the old name as an alias so that all cases should continue to run.