eilmer: extend flow Jacobian construction to 3D.
This commit extends the formation of our flow Jacobian to 3D. The implementation
has currently been verified for 1st and 2nd order Inviscid Jacobians.

This commit also removes the magic number system used to access the flow
variables in the arrays and matrices used for the adjoint and precondition
code. A similar approach to the steady-state solver has been applied to
generalise the code for 2D/3D laminar/turbulent options. Note however,
that we currently can't reuse the same variables that the steady-state solver
uses, due to the adjoint code structure. For now I have defined replica
variables attached to the block objects, in the near future I will refactor
the adjoint code so that we may directly use the same set of variables
employed in the steady-state solver.