kinetics: Allow electronically specific kinetics to be initialised with any general grouping strategy.
The electronically specific state solver previously was only able to be initialised to react between
predefined groups. This restricted the capacity to compare binning techniques easily as modifications
to the file structure were required. A python script now takes the reaction rates of individual levels
of nitrogen and oxygen, and their spectral energy data, and creates files based on a custom binning
strategy. These files contain the reaction rates between different binned groups, while the spectral
data is no longer required by the solver.

Examples of these files were included for the test version builds, but the python code to generate
these files has not been added to the commit. Old data files were removed.

Small changes in the gas area were made to account for the slightly varied format supplied by the
new files.

More significant changes were made in the kinetics area. init_thermochemical_reactor.d and the lua
wrapper were edited to initialise the electronically specific kinetics model with the new reaction
rate files. The kinetics file was edited to be initialised with the reaction rate files. This
required the creation of the rate fit parameter array to be created in the kinetics model
and passed to the state solver upon initialisation. Large blocks of now redundant code were
removed from the state solver.