Eilmer: introduce version labels: multi_species_gas, multi_T_gas, MHD, komega.
With these version labels, we will selectively build code to have or not have
each particular capability.  The intention is to reduce the amount of storage
required for all calculations.  We don't want large, ideal gas calculations
having to pay the storage and run-time cost of our full physical models.

By default, the code will be build to include all capabilities.  You may choose
to omit a certain capability by specifying the makefile option to be 0.
For example: make KOMEGA=0 install

This commit is the first of a few that will propagate the versions throughout
the code modules. So far, only the ConservedQuantities class has been adjusted.

Just accepting the default opt-in for all capabilities should have the code
build and behave as it did before.