Eilmer: first pass at putting version(xxxx) into fvcell.d
This is enough to get the compiler to not give errors on processing fvcell.d
when the capabilities are turned off.

Also cleaned up a bit of the selection code for the signal speeds where the MHD
bits had been added a while back.  Of course, this messes with the time-step
estimation a little and now some of the time-step checks fail... sigh.