Eilmer: clean up the automated tests, mainly number of steps has improved.
The decrease in the number of steps taken was a result of the fix to the signal
speed estimation code made in rev 2139.

The increase in the tolerance on the oxygen mass fraction for Bittker duct is to
accommodate a further small shift in calculated value that was aready close to
the specified tolerance limit.