gas: Restructuring of electronically specific gas model to incorporate the two temperature air thermodynamics and chemistry model.
This is the first commit in attempting to restructure the electronically specific gas model to properly capture other gas chemistry.
Much of the two temperature air code has been copied across, and implemented by specifying a macro_Q gas state, which
refers to the macrospecies of the gas without any electronically specific species. This allows for the same gas chemistry reactions
to take place, and will require an update for the true gas state Q, which will include individual species for electronically specific
gas species. The gas model must be implemented in a specific order, with the electronically specific states of atomic nitrogen
and oxygen at the start, followed by other gas species.

The LUA file to be passed to initialise the gas model will simply include two filenames, one with the information for the macro gas
species, and another for the information on the electronically specific states. The test versions of the gas and kinetics model
simply return 0 until further restructuring is done.