eilmer: added pseudo-time-stepping to adjoint solver.
This commit adds pseudo-time-stepping to the adjoint solver. By having
a pseduo-time-step we are able to improve the L.H.S. Jacobian matrices
diagonal dominance - this is intended to improve convergence of the
adjoint system for problems where solving the system directly with
GMRES stalls.

The formulation that is implemented is taken from Nielsen's work. The
code currently has fixed inner tolerance and intial CFL. These will
become user defined in a future revision. The implementation has been
 succesfully tested on a simple laminar ramp (coarse mesh) simulation.