kinetics: Disable high-temp correction to L-T relaxation time in air 2-T model
The high-temperature correction to the relaxation time calculation has been
causing some grief. The grief is that it causes a large delay in the approach
to thermal equilibrium for some 2-T air calculations. Gnoffo et al note that
this model is a little difficult to implement in a full flowfield. The high-
temperature correction strictly applies along a streamline and uses the
post-shock temperature for that streamline. This is difficult to implement
for all of the individual streamlines in a flow field, so an "average" post-shock
temperature is assumed in the model.

I suspect I have chosen a poor average and the literature does not describe
what choice they used for average post-shock temperature.

Presently, I have disabled this correction and things seem to behave reasonably.
I will need to do some deeper study on this model. I don't think we had this
available in eilmer3 and those calculations seemed to validate well against
results in the literature. Maybe this correction is not really necessary.