kinetics: Electronically specific kinetics now fully makes use of the chemistry and energy exchange of the two temperature model.
This commit seems to complete changes to the electronically specific kinetics to allow for most of the chemistry and energy
exchange to be handled by the two temperature model. The electronic modes of nitrogen and oxygen are then dealt with separately.
The current commit still requires an update to the energy after the electronic step, and for the macro species chemistry
to be altered to not consider reaction ionising N or O.