Eilmer: allow the user's supervisory function to override the time step.
On any time step, the user-defined supervisory function atTimestepStart
may set a Lua global variable dt_override to a positive value to limit the
allowable step size.  This will trigger a time-step check and dt_allow
will be limited by the value of dt_override.  If you do not wish to limit
dt_allow at any point, set the value of the Lua global variable dt_override
to nil, a negative number or zero.

Initially, dt_override is initialized to zero, so it will not be active unless
the user explicitly changes the Lua global variable to a positive number.