eilmer: write-out frozen limiter values for unstructured solver & use them in adjoint solver.
The adjoint system of equations requires the residuals to be dropped to machine precision. To accomplish this,
it is sometimes necessary to freeze the unstructured limiters in the flow solver to prevent limiter ringing/buzz.
When the limiters are frozen, the adjoint solver should use the frozen values.

This update allows the frozen limiter values to be written out to a file called frozen_limiter_values.dat, this file
is read by the adjoint solver if the flag read_frozen_limiter_values_from_file is set to true in the sscOptions set
in the lua input script.