eilmer: removed application of diffuseWallBCsOnInit from steadystate_solver.d
After the addition of diffuseWallBCsOnInit to the unsteady solver in r2182,
when running the steady-state solver we apply the diffusion process twice (once
in the shared initialisation code found in simcore.d and once in the specific
steady-state solver initialisation code).

This update removes the application of diffuseWallBCsOnInit in the steady-state
solver initialisation, i.e. relying on the application in simcore.d