kinetics: Decouple species identification from mass and charge balance.
We wish to handle some of the isomer species common in hydrocarbon
combustion chemistry. The symbols used in the species names do not
always directly map to elemental composition. Instead, we will now
rely on the elemental composition that is stored as the atomicConstituents
entry in the species database. This information is used to check that
elemental mass balances on either side of a chemical reaction.

Similarly, we will pull the species charge value directly from the database.

In the earlier version, I tried to glean this information from the species
name itself. This involved some complicated grammar parsing with LPEG.
This complication is now removed thanks to the new decoupled approach.