kinetics: Make chemkin2eilmer more tolerant of "dodgy" chemkin files.
Some sed processing has been added to clean up chemkin files before
converting to eilmer-style reaction input. I guess "dodgy" is a matter
of perspective. Certainly, some of the allowed input has ambiguities
that are difficult to parse.

Also, I plain do not like the use of 'AR' as a symbol for Argon. We already
have the perfectly acceptable and standard symbol of 'Ar' in our local
database. I decided to use that.