eilmer: implemented proper block-block boundary handling for 2nd order Jacobian.
The 2nd order (distributed) Jacobian, required for the adjoint solver, requires an
extra layer of ghost cells to correctly compute all of the off-block effects.

The new code essentially re-creates part of the neighbour blocks mesh in order to
get the necessary Jacobian entries. It does so by collecting the nearest-neighbours
of the mapped_cells (and their interfaces).

The new additions results in a block transparent adjoint solver for inviscid flows. Due
to the structuring of how we apply boundary conditions, the adjoint solver isn't block
transparent for vsicous flows, however the inaccuracies are confined to sections of the
domain where a block-block boundary condition attachs to a fluid domain boundary - as
such the accuracy is believed to be satisfactory.