kinetics: Bug fix in prep-chem to handle extraneous species in chemistry file.
Sometimes we like to list all the reactions in a scheme but only
use a subset of the model at run time. An example is a full 11-species
air chemistry model, but we'd only like to run it with 5 species.
When we process the chemistry file, the processor only has access to the
data for those 5 species so it balked on reactions that contained
"extaneous" species.

Previously, this worked because we broke apart the species tokens and
counted mass symbolically. Recently though we switched to using the
atomicConstituents data in the species database. So now we can't process
reactions with extraneous species because we have not loaded their data.

The solution is to change the processing order. The order is now:
1. Gather up all reactions, but leave as unprocessed tables.
2. Filter out unwanted reactions with "select" function filters.
3. Validate the "active" reactions.

(Previously, we had gathered and validated in one step.)