gasdyn: Remove default action to compute wedge case in Billig correlation
Our implementation always computed the shock angle of the flow
for a wedge, and then, the shock angle for a cone if needed.

However, as the Mach number increases, the shock angle for a wedge becomes
detached before a cone does. This meant that the Billig calculation couldn't
proceed. The fix is to compute one or the other depending on whether
a wedge or cone geometry is requested.