kinetics: Update prep-chem for two-temperature air model.
This is a bug fix and fall-out related to the change to using
atomicConstituents as a means to do mass balance check.
This information was not available in the two-temperature air
gas model file. The solution is to load all species information in
the prep-chem program and then pick from that giant database
as to what is needed.