Eilmer: Bring the limiter into play for the linear-interpolation
and, for extrema-clipping, just avoid doing the linear extrapolation because
that will have to introduce a new extreme value, if it results in anything
other than a copy of the cell data.

Loosen the tolerance of the shock-stand-off distance for the Sawada sphere case.
A width of two cells corresponds to a 14% change in stand-off distance and the
solution appears pretty chunky near the shock.