eilmer: add in domain mass balance when reporting residuals.
For complex flows the residuals tend to stall early, and provide
little information regarding whether a steady-state has been achieved.

This commit adds in a domain mass balance value that is reported when
the --report-residuals command line flag is used. This value essentially
represents the mass residual of the entire domain i.e. mass flowing in
minus mass flowing out.

Currently only available for the explicit version of the code.