Eilmer: user-defined ghost-cell and interface functions may return empty tables.
Empty tables indicate that they do not wish to overwrite the ghost-cell data
that was previously put in place by a boundary-data exchange.
If the returned tables are not empty, the action is to extract all of the flow
state data from those tables.  This is consistent with the old behaviour.