Eilmer: wrap the writing of the progress file in try-catch blocks.
There is a small possibility that the e4monitor process might try to read the
progress file at the same time as the main simulation goes to write a new value.
If this causes the write to fail, we ignore that failure and move on.
There will/should be another time-step shortly.