Eilmer: an experiment on time_derivatives function
...in which we restructure the code to put the loop over the cell's faces on
the outside.  This is an attempt to get better data locality by dealing with
all calculations for an interface in the body of the loop.  The trade-off,
of course, is that there is now more code branching within the loop, at least
for the most general build of the code.