solid: added a test implementation for least-squares gradient approximation.
The solid domain solver utilised the divergence theorem to approximate the
spatial gradients required in the temperature field update stage.

In preparation for extending our CHT implementation to 3D, a least-squares gradient approximation
has been implemented.
Currently, the implementation is hard-coded for 2D as a proof of concept.
The implementation achieves the expected spatial order of accuracy of 2 - determined from a CHT MMS test.

The solid domain solver will now use either divergence or least-squares to approximate
the spatial gradients depending on the spatial_deriv_calc config setting.