eilmer-cht: extended conjugate heat transfer solver to 3D.
These additions extend the conjugate heat transfer solver to 3D.
Currently, the implementation has been tested for a single fluid block
and single solid block configuration using a 3D MMS test case with no
variation in the z-direction. On a family of two refined grids, the
spatial order of accuracy is approximately 2 for this test case.