eilmer: ported over the Haselbacher viscous gradient method to the structured solver.
The augmented-face face-tangent method (locally refered to as the Haselbacher
method) has been ported from the unstructured solver to the structured solver.

To use this method, set the spatial derivative location to cells when using
the least squares gradient calculation procedure.

This method allows the structured solver to retain 2nd order spatial accuracy
everywhere in the domain when solving viscous simulations on domain
decomposed problems (i.e. multi-block simulations).

Note that for this method, we need the ghost cells to have the is_interior_to_domain
flag set to true.