gas: Fixed bug in 2-T air model related to electron contribution to energy.
I also took the opportunity to be more pedantic in the internal method
names. There are two lumped modes of energy in the 2-T air model:

1. transrotational; and
2. vibroelectronic

I used a more explicit name for the vibroelectronic mode. In this way,
I'll remember that the atomic species have contibutions in that mode
because they can store electronic energy if sufficiently excited.