gas: Use _species_ids array to identify electrons in 2-T air model.
We allow the user to declare the species in a 2-T air model
in any order. This is so we don't rewrite the species order behind
the scenes on them and lead to surprises in the output.

However, that means we need to handle the mapping from user-supplied
species order to what we use internally. This is done using the
_species_ids array. I had not been using this to identify the
index for the electron. Rather, my bug was to assume that the
electron was always the last item declared.