gas: Increase temperature bracket for first-pass Newton solve for T
In the Newton method with safeguards, an initial temperature bracket
is given for the search. My choice of 100 K for the bracket was too aggressive.
The consequence of this is that most iterative solve attempts failed, and so a very
expensive search from 20 - 10000 K as the fall-back was being used.

The new initial bracket of 1000 K seems to do much better. Most searches
pass with this value. What we are saying is that between flow updates,
we don't expect the temperature to change by more than 1000 K *frequently*.