Gas: make a place to put Nick's equilibrium chemistry functions.
This introduces a new gas model, which can be selected in the input file as:
model = 'ThermallyPerfectGasEquilibrium'
It is a derivative of ThermallyPerfectGas and has all of the same needs and
provides all of the same services.  The new class name is used to select
the equilibrium update module when the thermochemical reactor is constructed.