Gas: for the scripting-language wrappers, update sound speed when updating thermo.
Having the sound speed updated when other thermo properties are updated will be
less surprising when using the gas models from the Lua, Python and Ruby
scripting languages.

The core D-Language code will continue to require an explicit call to update
the sound speed because it may be an expensive computation for some of
the multi-species, multi-temperature gas models.  The flow solver doesn't
often need the sound speed updated when the other thermo properties are updated.